worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "aio-pika>=9.6.2",
    "asyncpg>=0.31.0",
    "cohere>=5.20.7",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "langchain-text-splitters>=1.1.1",
    "minio>=7.2.20",
    "numpy>=2.4.3",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
    "unstructured[pdf,docx]>=0.21.5",
    "pypdf>=6.9.2",
    "pandas>=3.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]
